Annotate heavy-light chain pairing

annotatePairing checks the repertoire table to annotate heavy-light chain pairing. This function classifies for a given cell (i.e. a given cell barcode) whether it is a singlet/doublet etc. based on the number of heavy (H) and light (L) chain sequences observed in the data

Usage

annotatePairing(tb_list, c_gene_column = "c_gene")

Arguments

tb_list

list of data.frame holding VDJ data. Each element corresponds to subset of sequences bearing one specific cell barcode

c_gene_column

Column name indicating where the isotype/light chain type information is stored in the data.frames in tb_list.

Value

Any one of 'singlet', 'BCR_doublet', 'multi_LC_same_class', 'multi_LC_diff_class' or 'multi_HC', according to the rules stated in the 'Details' section.

Details

The function assumes input of sequences all belonging to the same cell (i.e. bearing the same cell barcode). Based on the number of heavy and light chain sequences associated with a given cell barcode they are classified into one of the following categories:

BCR_doublets

More than 1 H and More than 1 L per cell barcode

multi_LC_diff_class

Exactly 1 H and more than 1 L per cell barcode, both kappa and lambda light chains are found

multi_LC_same_class

Exactly 1 H and more than 1 L per cell barcode, light chains are either all kappa or all lambda

multi_HC

More than 1 H and exacly 1 L per cell barcode

singlet

Exactly 1 H and/or exactly 1 L per cell barcode