Visualise stationary distribution of isotypes in the data

plotStationaryDistribution parse data from fitTPT() to plot a bar-plot of isotypes in the given data. The stationary distribution represents the equilibrium distribution of cells harbouring each isotype, taking the underlying isotype-switching dynamics into consideration.

Usage

plotStationaryDistribution(
  TPTObj,
  SeuratObj,
  ighc_count_assay_name = "IGHC",
  return_plot = TRUE
)

Arguments

TPTObj

List of TPT results, output from the fitTPT() function.

SeuratObj

Seurat object

ighc_count_assay_name

name of assay in SeuratObj which holds the IgH productive/sterile transcript counts. (Default: "IGHC")

return_plot

Should the CSR transition plot be returned? If FALSE, a named list of stationary_distribution and flux will be returned which contains the data frames to be visualised in this plot. (Default: TRUE)

Value

A ggplot2 object of the bar-plot of stationary distribution with 95

Details

plotStationaryDistribution parses data from the fitTPT() function for visualising the stationary distribution of isotypes. The resulting plot is a bar plot of stationary distribution, with error-bars obtained from bootstrapping (i.e. sampling with replacement) cells harbouring each isotype. The error-bar shown are 95